backend/requirements.txt
django-4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.2.0
openai==1.12.0

